pyproject.toml
[tool.poetry]
name = "instrumentdb"
version = "1.1.0"
description = ""
authors = ["Maurizio Tomasi <ziotom78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.10"
django = "^3.2.20"
markdown = "^3.2.1"
pygments = "^2.15.0"
djangorestframework = "^3.13.1"
django-filter = "^2.2.0"
pyyaml = "^6.0.0"
django-mptt = "^0.11.0"
gitpython = "^3.1.30"
envparse = "^0.2.0"
sphinx = "^5.1.1"
django-sslserver = "^0.22"
Pillow = "^9.3.0"
pytz = "^2022.1"
coverage = "^7.2.7"
flake8 = "^6.0.0"
black = "^23.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
manage = "manage:main"

[build-system]
requires = ["poetry>=1.3.1"]
build-backend = "poetry.masonry.api"
